fix: redraw rows after edit
